pyproject.toml
[project]
name = "datapizza-rag"
version = "0.1.0"
description = "Simple RAG MVP for CSV QA with PDF context."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "faiss-cpu>=1.7.4",
  "numpy>=1.24.0",
  "pypdf>=4.0.0",
  "python-dotenv>=1.0.0",
  "requests>=2.31.0",
  "sentence-transformers>=2.7.0"
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = [
  "src",
  "src.utils",
  "src.ingest",
  "src.rag",
  "src.eval"
]
